requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap4==2024.10
cryptography==44.0.2
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.2
django-braces==1.17.0
django-cleanup==8.0.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
weasyprint==60.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
pyzmq==26.3.0
qrcode==8.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
yarg==0.1.9
